backend/requirements.txt
# Core
django==6.0.4
django-tenants==3.10.1
psycopg[binary]==3.2.6

# REST API
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
django-filter==25.1
drf-spectacular==0.28.0

# Background tasks
celery==5.5.0
redis==5.3.0
django-celery-beat==2.9.0

# Email
django-anymail==13.1

# Utilities
python-decouple==3.8
Pillow==11.2.1
gunicorn==23.0.0
openpyxl==3.1.5
reportlab==4.4.0
